fix: Rename unit test for group and access construction
